services/worker/requirements.txt
arq>=0.26.0,<1.0
redis>=5.0,<6.0
boto3>=1.35.0,<2.0
python-dotenv>=1.0,<2.0
cadquery>=2.4.0,<3.0
mujoco>=3.0.0,<4.0
imageio>=2.34.0,<3.0
imageio-ffmpeg>=0.4.0,<1.0
